fix: Move formatSearchList in viewmodel
